CheatSheet/agent/requirements.txt
langchain==0.3.26
langgraph==0.4.10
langsmith==0.4.4
openai>=1.68.2,<2.0.0
fastapi>=0.115.5,<1.0.0
uvicorn>=0.29.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0 
langgraph-cli[inmem]==0.3.3
langchain-openai>=0.0.1
aiohttp>=3.9.0,<4.0.0
sse-starlette>=2.0.0,<3.0.0
redis>=5.0.0,<6.0.0
PyJWT>=2.8.0,<3.0.0
pydantic>=2.0.0,<3.0.0
